requirements.txt
# Core API & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
requests==2.31.0
httpx==0.27.0 
# Explicitly pin httpx to avoid conflicts with Groq/OpenAI clients

# AI & ML
torch>=2.0.0
sentence-transformers==2.3.1
huggingface_hub[cli,hf_transfer]
scikit-learn==1.4.0
numpy>=1.24.0
transformers==4.37.2

# LLM Client
groq==0.4.0

# Legacy & Utils
streamlit==1.31.0
pytest==8.0.0
python-multipart==0.0.9
